content/media/video/test/test_bug465498.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: Bug 465498 - Seeking after playback ended</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=465498">Mozilla Bug 465498</a>
<audio id='v'
       onloadedmetadata='return startTest();'
       onended='return playbackEnded();'
       onseeked='return endTest();'>
  <source type='audio/x-wav' src='r11025_s16_c1.wav'>
</audio>
<pre id="test">
<script class="testbody" type="text/javascript">
var v = document.getElementById('v');
var haveSeeked = false;

function startTest() {
  v.play();
  return false;
}

function playbackEnded() {
  ok(v.currentTime >= 0.9 && v.currentTime <= 1.1,
     "Checking currentTime at end: " + v.currentTime);
  ok(v.ended, "Checking playback has ended");

  v.currentTime = 0;
  haveSeeked = true;

  return false;
}

function endTest() {
  if (!haveSeeked)
    return false;

  ok(v.currentTime == 0, "Checking currentTime after seek: " + v.currentTime);
  ok(!v.ended, "Checking ended is false");
  SimpleTest.finish();
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
